requirements.txt
altgraph==0.17.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
elasticsearch7==7.17.6
future==0.18.2
idna==3.4
kafka-python==2.0.2
macholib==1.15.2
pdf2image==1.16.0
Pillow==9.2.0
requests==2.28.1
six==1.15.0
soupsieve==2.3.2.post1
urllib3==1.26.12
